Code Refactor - created methods for date conversions
Removed the need to query the database a second time to get all of the
comments for issue and comment date conversions.  Issue and Comments
date fields are now converted during the intial pull from the API and
dates are converted before they are pushed into the database for the
first time.